Core mechanics coded
Got the core mechanics of the game to work. You have to press the keys in the right order to win each round. Each round, there'll be more and more keys that you have to press in the right order to win the round. They list of keys to press is also displayed on the left side of the screen.